Addresses Argyris's comments
